move validators and metas to their respective modules
