+++ kmk/modules/tapdance.py
@@ -1,6 +1,5 @@
 from kmk.keys import KC, make_argumented_key
-from kmk.modules.holdtap import ActivationType, HoldTap
-from kmk.types import HoldTapKeyMeta
+from kmk.modules.holdtap import ActivationType, HoldTap, HoldTapKeyMeta
 
 
 class TapDanceKeyMeta:
